requirements.txt
toml
Pillow
opencv-python
opencv-contrib-python
scipy
torch==2.5.1
torchaudio==2.5.1
decord
pydub
tqdm
matplotlib
linmult
pyyaml
prettytable
lightning
seaborn